feat(client): updating engine workspace WIP
